move steps to module; cleanup
