+++ frontend/src/reports/ReportProviderSteps.ts
@@ -1,0 +1,40 @@
+import { StepData } from "../pages/ExploreData/TableOfContentsStepper";
+
+export const reportProviderSteps: StepData[] = [
+  {
+    label: "Population",
+    hashId: "population",
+  },
+  {
+    label: "Rate Map",
+    hashId: "map",
+  },
+  {
+    label: "Rate Chart",
+    hashId: "bar",
+  },
+  {
+    label: "Unknown Share Map",
+    hashId: "unknowns",
+  },
+  {
+    label: "Share Chart",
+    hashId: "disparity",
+  },
+  {
+    label: "Data Table",
+    hashId: "table",
+  },
+  {
+    label: "Age-Adjusted Ratios",
+    hashId: "age-adjusted",
+  },
+  {
+    label: "Definitions",
+    hashId: "definitions",
+  },
+  {
+    label: "What Data Are Missing?",
+    hashId: "missingDataInfo",
+  },
+];
